parse_rdesc: rename HidItem and HidInputItem
HidItem is the raw item in the report descriptor. It doesn't contain more
than the current value, the size and the type.
Rename it to HidRdescItem.

HidInputItem is the parsed field which matches to one field in the
input/output/feature report. We should enhance it with more information
(applications, etc..) and subclass it for the 3 different types.
Rename it into HidField.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>