tests: do not try to open input nodes when the hid device disappears
We usually stop polling the uhid and udev events as soon as we find
the one device we need. If a secondary device appears after, the udev
event will stay queued until we process the events again, when we check
for the devices removal.

Make sure we do not trigger input nodes creation in those cases.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>