tests: add conftest.py for pytest environments
When we run pytest with the xdist plugin to have parallel execution,
trying to setup and delete the same udev rule multiple times is not
a good idea.

Implement a pytest fixture with a specific name per session so we can
be sure to have the rule in place until the end of the session.

This allows to now run 'sudo pytest-3 -n 4' to have 4 parallel executions
and divide the time spent by 2.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>