Fix hid-replay.py script
parse_rdesc() expects to have the length in the first field of the given
string. We do not really care about its value, but hid-replay is
smart enough to split the length out of the rest and compare it
to the result.

Make the uhid device interface saner where we do not need to provide
the length, just the data.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>