parse_hid: prettify the alignment of the following rows in a dump
If there are some non recurrent fields at the beginning of the report,
the previous code would mess up the indentation, making it even more
difficult to read.

Fixes that by looking where in the previous part we are matching the
first field of the second row and keep this as the indent value.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>