tests: mt: add Elan touchpad weird behavior
Some Elan touchpads declare 2 InputMode feature in the configuration
collection. The first one is handled properly (as per the spec) while
the second one leads to some unexpected behavior.

In the 3045 case, the touchpad doesn't send the physical clicks while
the finger is in the lower right area.

Make sure the configuration gets busted in such case to prevent future
mis-behavior.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>