uhid: make the ready flag private
No need to do keep this as a public flag, all we do ever is drop the event if
the device is about to be removed.

The only use-case for this would be if a subclass cares about the udev device
removed event, but that's not what we have atm.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>