hidraw: make read_events a bit more useful
Returning the bytes is unlikely to excite anyone, so let's return something
more useful: a tuple that points to the events we just read from the device.

Also: in the presumably rare case where we end up with 4096 bytes of data to
read - keep on reading until we have pulled all pending events.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>