test: make sure we also pass the tests on a non high res wheel kernel
If the kernel doesn't export high res wheel udev bits, we can deduce it
doesn't support high resolution wheels. And based on that, we can decide
skipping or not the high resoltuion tests.

We need to 'emulate' the firmware in so that all wheel mice have a
resolution multiplier of 1 by default, and it gets changed when we receive
the feature from the kernel.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>