record: fix when there is no hidraw node provided
When we do not have any hidraw node, we should provide a list of
the existing nodes and let the user decide.

Using '+' enforces to have a hidraw node, as stated in the doc:
'+'. Just like '*', all command-line args present are gathered into
a list. Additionally, an error message will be generated if there
wasn’t at least one command-line argument present.

So we prefer having '*'

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>