tests: add a custom test for reallocation in the kernel collection stack
A kernel bug had some internal pointers expire after reallocation. For an
explanation, see kernel patch
   HID: core: replace the collection tree pointers with indices

This test triggers that reallocation but it cannot reliably trigger the actual
bug (a crash). So it only works as a negative test, i.e. if it ever fails we
know why but succeeding does not show absence of the bug.

The test device here is a copy/paste device with several (more than 16)
collections added. kernel printks verified that the device triggers the
previously faulty reallocation of the faulty array and the Resolution
Multiplier means we actually use the parent pointers. But as said above, it
cannot be reliably made to crash, so this test is only useful when it fails.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>