tests: check for the kernel taint after each test
This will allow to catch the first oops or warning. And this is
notably useful for checking if the reallocation of the kernel collection
stack is correct.

However, once a kernel is tainted, we can't reset the flag short of
rebooting the kernel. So this means we can only catch the first kernel
warning.

Note: enable KASAN to have a reliable detection of failure in the
previously mentioned test.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>